+++ domain/src/main/java/com/example/yourfinance/domain/model/entity/Payment.kt
@@ -4,13 +4,14 @@ import com.example.yourfinance.domain.model.Title
 import com.example.yourfinance.domain.model.Transaction
 import com.example.yourfinance.domain.model.TransactionType
 import com.example.yourfinance.domain.model.entity.category.ICategoryData
+import java.math.BigDecimal
 import java.time.LocalDate
 import java.time.LocalTime
 
 
 data class Payment(
     override var type: TransactionType,
-    override var balance: Double,
+    override var balance: BigDecimal,
     var moneyAccount: MoneyAccount,
     var category : ICategoryData,
     private var _note: Title,
