+++ presentation/src/main/java/com/example/yourfinance/presentation/ui/adapter/list_item/TransactionListItem.kt
@@ -1,10 +1,11 @@
 package com.example.yourfinance.presentation.ui.adapter.list_item
 
 import com.example.yourfinance.domain.model.Transaction
+import java.math.BigDecimal
 import java.time.LocalDate
 
 sealed class TransactionListItem {
-    data class Header(val date: LocalDate, val balance: Double) : TransactionListItem()
+    data class Header(val date: LocalDate, val balance: BigDecimal) : TransactionListItem()
     data class TransactionItem(val transaction: Transaction) : TransactionListItem()
     object EmptyItem : TransactionListItem()
 }
